Implemented enforce availability feature
